+++ Config/DefaultEngine.ini
@@ -106,8 +106,8 @@ ManualIPAddress=
 +Profiles=(Name="Ragdoll",CollisionEnabled=QueryAndPhysics,bCanModify=False,ObjectTypeName="PhysicsBody",CustomResponses=((Channel="Pawn",Response=ECR_Ignore),(Channel="Visibility",Response=ECR_Ignore)),HelpMessage="Simulating Skeletal Mesh Component. All other channels will be set to default.")
 +Profiles=(Name="Vehicle",CollisionEnabled=QueryAndPhysics,bCanModify=False,ObjectTypeName="Vehicle",CustomResponses=,HelpMessage="Vehicle object that blocks Vehicle, WorldStatic, and WorldDynamic. All other channels will be set to default.")
 +Profiles=(Name="UI",CollisionEnabled=QueryOnly,bCanModify=False,ObjectTypeName="WorldDynamic",CustomResponses=((Channel="WorldStatic",Response=ECR_Overlap),(Channel="Pawn",Response=ECR_Overlap),(Channel="Visibility"),(Channel="WorldDynamic",Response=ECR_Overlap),(Channel="Camera",Response=ECR_Overlap),(Channel="PhysicsBody",Response=ECR_Overlap),(Channel="Vehicle",Response=ECR_Overlap),(Channel="Destructible",Response=ECR_Overlap)),HelpMessage="WorldStatic object that overlaps all actors by default. All new custom channels will use its own default response. ")
-+Profiles=(Name="Weapon",CollisionEnabled=QueryAndPhysics,bCanModify=True,ObjectTypeName="Weapon",CustomResponses=((Channel="WorldStatic",Response=ECR_Overlap),(Channel="WorldDynamic",Response=ECR_Overlap),(Channel="Pawn",Response=ECR_Overlap),(Channel="Visibility",Response=ECR_Overlap),(Channel="Camera",Response=ECR_Overlap),(Channel="PhysicsBody",Response=ECR_Overlap),(Channel="Vehicle",Response=ECR_Overlap),(Channel="Destructible",Response=ECR_Overlap),(Channel="Character",Response=ECR_Ignore),(Channel="Weapon",Response=ECR_Ignore)),HelpMessage="Needs description")
-+Profiles=(Name="Character",CollisionEnabled=QueryOnly,bCanModify=True,ObjectTypeName="Character",CustomResponses=((Channel="Visibility",Response=ECR_Ignore),(Channel="Weapon",Response=ECR_Ignore)),HelpMessage="Needs description")
++Profiles=(Name="Weapon",CollisionEnabled=QueryAndPhysics,bCanModify=True,ObjectTypeName="Weapon",CustomResponses=((Channel="WorldStatic",Response=ECR_Overlap),(Channel="WorldDynamic",Response=ECR_Overlap),(Channel="Pawn",Response=ECR_Overlap),(Channel="Visibility",Response=ECR_Overlap),(Channel="Camera",Response=ECR_Overlap),(Channel="PhysicsBody",Response=ECR_Overlap),(Channel="Vehicle",Response=ECR_Overlap),(Channel="Destructible",Response=ECR_Overlap),(Channel="Character",Response=ECR_Ignore),(Channel="Weapon",Response=ECR_Ignore)),HelpMessage="Collision for Weapon")
++Profiles=(Name="Character",CollisionEnabled=QueryOnly,bCanModify=True,ObjectTypeName="Character",CustomResponses=((Channel="Visibility",Response=ECR_Ignore),(Channel="Weapon",Response=ECR_Ignore)),HelpMessage="Collision for the Character")
 +DefaultChannelResponses=(Channel=ECC_GameTraceChannel1,DefaultResponse=ECR_Block,bTraceType=False,bStaticObject=False,Name="Character")
 +DefaultChannelResponses=(Channel=ECC_GameTraceChannel2,DefaultResponse=ECR_Overlap,bTraceType=False,bStaticObject=False,Name="Weapon")
 +EditProfiles=(Name="Pawn",CustomResponses=((Channel="Visibility",Response=ECR_Ignore)))
