+++ Source/InfiniteFighter/Character/IFCharacter.cpp
@@ -13,7 +13,7 @@
 #include "IFCharacterAnimInstance.h"
 #include "Components/WidgetComponent.h"
 #include "IFAxe.h"
-#include "IFAimWidget.h"
+#include "UI/IFAimWidget.h"
 #include "Components/CapsuleComponent.h"
 
 // Sets default values
