+++ Source/InfiniteFighter/InfiniteFighter.Build.cs
@@ -8,6 +8,8 @@ public class InfiniteFighter : ModuleRules
 	{
 		PCHUsage = PCHUsageMode.UseExplicitOrSharedPCHs;
 
+		PublicIncludePaths.AddRange(new string[] { "InfiniteFighter" });
+
 		PublicDependencyModuleNames.AddRange(new string[] { "Core", "CoreUObject", "Engine", "InputCore", "HeadMountedDisplay", "EnhancedInput", "AnimGraphRuntime", "CommonInput", "UMG" });
 	}
 }
