--- Source/InfiniteFighter/InfiniteFighterGameMode.h
@@ -1,19 +1,0 @@
-// Copyright Epic Games, Inc. All Rights Reserved.
-
-#pragma once
-
-#include "CoreMinimal.h"
-#include "GameFramework/GameModeBase.h"
-#include "InfiniteFighterGameMode.generated.h"
-
-UCLASS(minimalapi)
-class AInfiniteFighterGameMode : public AGameModeBase
-{
-	GENERATED_BODY()
-
-public:
-	AInfiniteFighterGameMode();
-};
-
-
-
